Added in animation
Added in the transition styles for the game navigation bar. Renamed the files index.html and paa_animate.css. Added in the bat, fox, and butterfly for the animation. Added in the animation keyframes that controls the animation. Fixed a bug where the bat sprite would get clipped every once in awhile. Added in a few comments.